backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
sentence-transformers
chromadb
flask
flask-cors
flask-sqlalchemy
werkzeug
python-dotenv
gunicorn